# .tmp_backend_openapi.yaml
openapi: 3.0.3
info:
  title: Loyalty Platform API
  version: 1.0.0
  description: Personalized loyalty + recommendations platform
paths:
...
  /api/me/next-offer:
    get:
      operationId: me_next_offer_retrieve
      description: Get active offer assignment or auto-assign next offer under budget/cooldown
        constraints.
      tags:
      - Offers
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
              examples:
                NextOffer(bundle-drivenExample):
                  value:
                    assignment_id: 4
                    offer:
                      id: 1
                      name: whosadik
                      type: discount
                      value: '2.00'
                      estimated_cost: '5.00'
                    target:
                      scope: product_id
                      value: 322
                      category: makeup
                      product_type: eyeshadow
                      picked_via: bundle
                      bundle_mode: cooccurrence
                      bundle_why:
                      - frequently purchased with product_id=330 (count=1)
                    reason:
                      segment: new_or_rare
                      picked_because: max(score) under eligibility + cooldown + budget
                        constraints
                      post_purchase:
                        categories:
                        - makeup
                        product_types:
                        - foundation
                  summary: Next offer (bundle-driven example)
          description: ''
  /api/me/offers:
    get:
      operationId: me_offers_retrieve
      tags:
      - me
      security:
      - cookieAuth: []
      responses:
        '200':
          description: No response body
...
  /api/offers/click:
    post:
      operationId: offers_click_create
      description: Record user click on active offer assignment (idempotent by request
        key).
      tags:
      - Offers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfferClickRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/OfferClickRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OfferClickRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/offers/redeem:
    post:
      operationId: offers_redeem_create
      description: |-
        Применяем оффер к транзакции и начисляем баллы в ledger.
        MVP: реализуем points_multiplier.
      tags:
      - Offers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedeemOfferRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RedeemOfferRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RedeemOfferRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedeemOfferResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: ''
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        ok:
          type: boolean
          default: false
        code:
          type: string
        message:
          type: string
        details:
          nullable: true
        request_id:
          type: string
          nullable: true
      required:
      - code
      - message
...
    OfferClickRequest:
      type: object
      properties:
        assignment_id:
          type: integer
        context: {}
      required:
      - assignment_id
...
    RedeemOfferRequest:
      type: object
      properties:
        assignment_id:
          type: integer
        transaction_id:
          type: integer
      required:
      - assignment_id
      - transaction_id
    RedeemOfferResponse:
      type: object
      properties:
        ok:
          type: boolean
        earned_points:
          type: integer
        new_balance:
          type: integer
        tier:
          type: string
          nullable: true
        discount_amount:
          type: string
      required:
      - discount_amount
      - earned_points
      - new_balance
      - ok
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid